# web/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditVerifyResponse": {
        "properties": {
          "first_broken_sequence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Broken Sequence"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid",
          "first_broken_sequence"
        ],
        "title": "AuditVerifyResponse",
        "type": "object"
      },
...
      "DriftMetric": {
        "properties": {
          "baseline": {
            "title": "Baseline"
          },
          "current": {
            "title": "Current"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "name",
          "kind",
          "status"
        ],
        "title": "DriftMetric",
        "type": "object"
      },
      "DriftReport": {
        "properties": {
          "alerts": {
            "items": {
              "type": "string"
            },
            "title": "Alerts",
            "type": "array"
          },
          "baseline_created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Created At"
          },
          "current": {
            "$ref": "#/components/schemas/DriftSnapshot"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "has_baseline": {
            "title": "Has Baseline",
            "type": "boolean"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/DriftMetric"
            },
            "title": "Metrics",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "has_baseline",
          "current",
          "metrics",
          "alerts",
          "status"
        ],
        "title": "DriftReport",
        "type": "object"
      },
      "DriftSnapshot": {
        "properties": {
          "categorical": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "integer"
              },
              "type": "object"
            },
            "title": "Categorical",
            "type": "object"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "histograms": {
            "additionalProperties": {
              "items": {
                "type": "integer"
              },
              "type": "array"
            },
            "title": "Histograms",
            "type": "object"
          },
          "samples": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Samples",
            "type": "object"
          },
          "scalars": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Scalars",
            "type": "object"
          },
          "versions": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Versions",
            "type": "object"
          },
          "window_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Start"
          }
        },
        "required": [
          "counts",
          "categorical",
          "histograms",
          "scalars",
          "versions"
        ],
        "title": "DriftSnapshot",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Ingestion, classification, extraction, validation, grounded Q&A with citations, human review, audit and evaluation for financial documents.",
    "title": "Financial Document Intelligence API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/audit/verify": {
      "get": {
        "description": "Recompute the audit hash chain to detect after-the-fact modification.",
        "operationId": "verify_audit_chain_audit_verify_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "cookie",
            "name": "docintel_demo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Docintel Demo"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditVerifyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Audit Chain",
        "tags": [
          "system"
        ]
      }
    },
...
    "/drift/report": {
      "get": {
        "description": "Compare current operational data against the stored drift baseline.",
        "operationId": "drift_report_drift_report_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "cookie",
            "name": "docintel_demo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Docintel Demo"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriftReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Drift Report",
        "tags": [
          "system"
        ]
      }
    },
...
  }
}